implement authorizations for delete and update

+++ app/Http/Controllers/QuestionController.php
@@ -53,7 +53,9 @@ class QuestionController extends Controller
      */
     public function show(Question $question)
     {
-        //
+        $question->increment('views');
+
+        return view('questions.show', compact('question'));
     }
 
     /**
@@ -64,6 +66,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access Denied!');
+        }
+        
         return view('questions.edit', compact('question'));
     }
 
@@ -76,6 +82,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access Denied!');
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', 'Your question has been upadted.');
@@ -89,6 +99,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, 'Access Denied!');
+        }
+        
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted.');
